Nvim: use nvim-cmp
We change the autocompletion engine from the deprecated nvim-compe to
nvim-cmp. This is a much better autocompletion library, supporting
completion even in command mode and is much more customisable, and seems
to be faster.

We also update the delimiter library (delimitMate) to a more modern
alternative and tweak a few other settings.